cerbero: Switch enums from OS_X to MACOS
Apple stopped using "OS X" in 2016, and 8 years later let's start
moving away from it. We need this for Swift support which uses
"macos", "ios" etc in the target triple instead of "osx" or "darwin".

Not replacing all occurrences of OSX, such as in relocation and
packaging because that would be spammy and we need to change those
anyway when we move to xcframework.