Differentiate requested from the installed files
Currently, the different "search" methods are not consistent
on what they do. Some of them don't validate on the file system
for the existence of the files, some do, and some others do
further processing of the files to find the actual requested file.
In the same situation, not finding files is some times considered
an error and some others isn't.

Split the search from the validation of on-disk files process.
Also generalize the criteria of error if a file is not found.

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1477>